dom/tests/mochitest/pointerlock/file_cursorPosEvents.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_cursorPosEvents.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">
    #fs > div { width:100px; height:100px; }
  </style>
</head>
<body onload="setup();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
  Mozilla Bug 633602</a>
  <p id="display"></p>
<div id="content">
  <div id="fs">
    <div id="one" style="background-color:Green;"></div>
    <div id="outer" style="background-color:Gray;">
    </div>
  </div>
  <pre id="test">
    <script type="application/javascript">
      /*
       * Test for Bug 633602
       * Test will check to make sure that the following mouse events are no
       * longer executed in pointer lock.
       *  - mouseover, mouseout, mouseenter, mouseleave
       */
      SimpleTest.waitForExplicitFinish();
      var moved = 0;
      var fs = document.getElementById('fs');
      var one = document.getElementById('one');
      var outer = document.getElementById('outer');
      var pointer = navigator.mozPointer;
      var mouseEnter = 0;
      var mouseLeave = 0;
      var mouseOver = 0;
      var mouseOut = 0;

      document.addEventListener('mozfullscreenchange', mouseCursorRequired,
        false);

      function mouseCursorRequired() {
        if (document.mozFullScreen && document.mozFullScreenElement === fs) {
          pointer.lock(fs, cursorActions, function () {
            ok(pointer.isLocked, "Pointer should lock.")
          });
        } 
      }

      function cursorTest() {
        if (moved === 0) {
          moved++;
          synthesizeMouseAtCenter(one, { type: 'mousemove' }, window);
        } else if (moved === 1) {
          moved++;
          synthesizeMouseAtCenter(outer, { type: 'mousemove' }, window);
        } else {
          is(mouseEnter, 0,
            "mouseenter should not be firing in Full Screen and Pointer Lock.");
          is(mouseOver, 0,
            "mouseover should not be firing in Full Screen and Pointer Lock.");
          is(mouseLeave, 0,
            "mouseleave should not be firing in Full Screen and Pointer Lock.");
          is(mouseOut, 0,
            "mouseout should not be firing in Full Screen and Pointer Lock.");
          pointer.unlock();
        }
      }

      function cursorActions() {
        fs.addEventListener('mousemove', cursorTest, false);
        synthesizeMouseAtCenter(outer, { type: 'mousemove' }, window);
      }

      function setup() {
        document.addEventListener('mozpointerlocklost', function() {
          document.mozCancelFullScreen();
          SimpleTest.finish();
        }, false);

        // Events that only exist because a mouse exists
        one.addEventListener('mouseout', function() {
          mouseOut = 1;
        } , false);

        one.addEventListener('mouseleave', function() {
          mouseLeave = 1;
        } , false);

        one.addEventListener('mouseover', function() {
          mouseOver = 1;
        } , false);

        one.addEventListener('mouseenter', function() {
          mouseEnter = 1;
          } , false);

        startTests();
      }

      function startTests() {
        SimpleTest.waitForFocus(function() {
          fs.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
